broke out self promo into slash and app
